Changed zlib crc type from int to unsigned long, as that's the type returned by zlib's crc32() and doing so reportedly resolves the CRC error reported in [forum:7a9097e557 | forum post 7a9097e557].
FossilOrigin-Name: 1bf6686e76252efda0abf7c64da8d5abb61c663187016ff77c4ac3f774988fc6

+++ src/gzip.c
@@ -31,7 +31,7 @@
 */
 struct gzip_state {
   int eState;           /* 0: idle   1: header  2: compressing */
-  int iCRC;             /* The checksum */
+  unsigned long iCRC;   /* The checksum */
   z_stream stream;      /* The working compressor */
   Blob out;             /* Results stored here */
 } gzip;
